Use a single global timer thread to schedule all time related tasks
This will ensure only one timer thread is created which according to the
documentation should be sufficient for handling thousands of tasks.

This also ensure the server can shut down properly when we close the
server window since these timer threads were running as non-daemons,
which meant that when we tried to close the server it kept going in the
background because of the timer threads. Now we have a single timer
thread that gets stopped when the server window is closed.